+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## [0.5.2]
+
+### Fix
+
+- Cannot see params & return value, for global functions (#136)
+
+## [0.5.1]
+
+### Fix
+
+- ignore "restrict" keyword in C function parameter pointer (#121)
+
+- Parameters not generated when using custom trigger (#119)
+
+- Function argument. Pointer to table is not generated. (#123)
+
+### Other
+
+- Update dev dependencies and increase min VS Code version (#131, #132)
+
 ## [0.5.0]
 
 ### Feature